requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
python-dotenv==1.0.0
pillow==11.3.0
aiofiles==23.2.1
jinja2==3.1.2
pandas==2.1.2
openpyxl==3.1.2
bcrypt==4.1.2
slowapi==0.1.9
email-validator==2.1.0
